% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonpower_bin.R
\name{Poissonpower.bin}
\alias{Poissonpower.bin}
\title{Sample size and power for Poisson regression.}
\usage{
Poissonpower.bin(alpha = NULL, power = NULL, exp.B0 = NULL,
  exp.B1 = NULL, muT = 1, phi = 1, R2 = 0, pi.x1 = NULL, N = NULL)
}
\arguments{
\item{alpha}{type I error rate. Can range from 0 to 1 (typically 0.05)}

\item{power}{1 - Pr(type II error) Can range from 0 to 1 (typically 0.80)}

\item{exp.B0}{Baseline rate. The response rate when all covariates have a value of 0}

\item{exp.B1}{Incidence rate ratio (IRR) comparing X=1 to X=0.}

\item{muT}{Mean exposure time}

\item{phi}{Measure of over-dispersion}

\item{R2}{The square of the multiple correlation coefficient when the covariate of interst is regressed on the other covariates.}

\item{pi.x1}{proportion of sample (N) with X1=1}

\item{N}{sample size}
}
\value{
Return one of the following parameters

\code{alpha}, \code{power}, \code{exp.B1}, \code{N}
}
\description{
Calculating power/sample size for simple Poisson regression with a binary predictor. The function solves for one of the following: alpha, power, incidence rate ratio, N.
}
\note{
The test is a two-sided test. For one-sided tests, double the significance level. For example, you can set alpha=0.10 to obtain one-sided test at 0.05 significance level.

\code{alpha}, \code{power}, \code{exp.B0}, \code{exp.B1}, \code{pi.x1}, \code{N} can be input as either single values or vectors. Only one parameter can be input as a vector.
}
\examples{
Poissonpower.bin(alpha=NULL, power=.80, exp.B0=.85, exp.B1=1.3, muT=1, phi=1,R2=0, pi.x1=.5, N=406)
Poissonpower.bin(alpha=NULL, power=.80, exp.B0=.85, exp.B1=1.3, muT=1, phi=1, R2=0, pi.x1=.5, N=406)
Poissonpower.bin(alpha=.05, power=.80, exp.B0=.85, exp.B1=1.3, muT=1, phi=1, R2=0, pi.x1=.5, N=NULL)
Poissonpower.bin(alpha=.05, power=.80, exp.B0=.85, exp.B1=NULL, muT=1, phi=1, R2=0, pi.x1=.5, N=406)
}
\references{
Signorini, D. 1991. Sample Size for Poisson Regression, Biometrika (1991), 78, 2, pages 446-450..
}
\author{
David Aaby <david.aaby@northwestern.edu>
}
